+++ utils/testng_utils.go
@@ -1,7 +1,7 @@
 package utils
 
 import (
-	"path/filepath"
+
 )
 
 type GetTestNGArgsStruct struct{
@@ -15,21 +15,21 @@ func GetTestNGArgs(c GetTestNGArgsStruct) GetTestNGArgsStruct {
 	c.EnvVarsFile = GetInputFromStdin(
 		GetInputFromStdinStruct{
 			Prompt: []string{"script where env vars are set for the app to run relative to workspace root"},
-			Default: filepath.Join(c.WorkspaceFolder,"ignore\\Local\\testng_e2e_shared.env"),
+			Default: JoinAndConvertPathToOSFormat(c.WorkspaceFolder,ConvertPathToOSFormat(".windmillcode\\Local\\testng_e2e_shared.env")),
 		},
 	)
 
 	c.TestNGFolder = GetInputFromStdin(
 		GetInputFromStdinStruct{
 			Prompt: []string{"testng app location"},
-			Default: filepath.Join(c.WorkspaceFolder,"apps\\testing\\testng"),
+			Default: JoinAndConvertPathToOSFormat(c.WorkspaceFolder,ConvertPathToOSFormat("apps\\testing\\SeleniumApp")),
 		},
 	)
 
 	c.SuiteFile = GetInputFromStdin(
 		GetInputFromStdinStruct{
 			Prompt: []string{"xml suite file needed for testng (this should be relative to the testng folder)"},
-			Default: filepath.Join("src\\test\\resources\\tests.xml"),
+			Default: ConvertPathToOSFormat("src\\test\\resources\\tests.xml"),
 		},
 	)